+++ CHANGELOG.md
@@ -7,6 +7,7 @@ ChangeLog
 
 - IMPROVEMENT: use a nested tab experience for a more clean control UI
 - IMPROVEMENT: store also Streams configuration in presets
+- IMPROVEMENT: make stream source configurable for Decal, Monitor, Plate and Hologram
 - BUGFIX: fix preset filters by adding Plate and Hologram
 
 2.1.2 (2024-09-16)